Update docs, fix "host" input bug, remove old junk, refactor DnD
