+++ src/sequences/sequences_list.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { selectSequence, addSequence } from "./actions";
 import { SequenceReducerState } from "./interfaces";
 import { Sequence } from "./interfaces";
-import { AuthToken } from "../auth/actions";
+import { AuthToken } from "../auth/interfaces";
 import { t } from "i18next";
 
 let buttonList = (dispatch: Function) => (seq: Sequence, index: number) => {
